* subversion/tests/svn_test_main.c   (do_test_num): Fix compile-time warning by using correct type.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1049328 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/tests/svn_test_main.c
@@ -260,7 +260,7 @@ do_test_num(const char *progname,
 
   if (msg)
     {
-      int len = strlen(msg);
+      size_t len = strlen(msg);
       if (len > 50)
         printf("WARNING: Test docstring exceeds 50 characters\n");
       if (msg[len - 1] == '.')
